+++ visual_genome/__init__.py
@@ -1,0 +1,3 @@
+
+VERSION_INFO = (1, 0, 0)
+__version__ = '.'.join(map(str, VERSION_INFO))
