+++ visual_genome/data/getImageData.sh
