+++ visual_genome/data/getQuestionAnswers.sh
