+++ visual_genome/data/getRegionDescriptions.sh
